requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
arrow==0.4.3
distribute==0.6.24
itsdangerous==0.24
python-dateutil==2.2
python-redmine==0.8.2
requests==2.3.0
six==1.7.3
wsgiref==0.1.2
